Auth-system-updated

+++ JavaScript/11-Web-API/2-Auth-System/2-Auth-LocalStorage/js/signup.js
@@ -1,3 +1,4 @@
+const signupForm = document.querySelector("#signup-form");
 const firstName = document.querySelector("#firstname");
 const lastName = document.querySelector("#lastname");
 const email = document.querySelector("#email");
@@ -6,19 +7,38 @@ const confirmPassword = document.querySelector("#confirm-password");
 const submitBtn = document.querySelector("#submit-btn");
 
 submitBtn.addEventListener("click", (e) => {
+  signupForm.classList.add("was-validated");
   e.preventDefault();
-  if (password.value === confirmPassword.value) {
-    const userDetails = {
-      id: Math.random().toString(16).split(".")[1],
-      firstName: firstName.value,
-      lastName: lastName.value,
-      email: email.value,
-      password: password.value,
-      confirmPassword: confirmPassword.value,
-    };
-    console.log(userDetails);
-    localStorage.setItem(userDetails.id, JSON.stringify(userDetails));
+  if (
+    firstName.value &&
+    lastName.value &&
+    email.value &&
+    password.value &&
+    confirmPassword.value
+  ) {
+    if (password.value === confirmPassword.value) {
+      const userDetails = {
+        id: Math.random().toString(16).split(".")[1],
+        firstName: firstName.value,
+        lastName: lastName.value,
+        email: email.value,
+        password: password.value,
+        confirmPassword: confirmPassword.value,
+      };
+      console.log(userDetails);
+      localStorage.setItem(userDetails.id, JSON.stringify(userDetails));
+      alert("User created successfully");
+      // window.location.reload();
+      firstName.value = "";
+      lastName.value = "";
+      email.value = "";
+      password.value = "";
+      confirmPassword.value = "";
+      signupForm.classList.remove("was-validated");
+    } else {
+      alert("password and confirm password must be same");
+    }
   } else {
-    alert("Password and Confirm Password not same");
+    alert("Fields should not be empty");
   }
 });
